Added BookService and added endpoints to LibraryController
In BookService added a method that allows the client to checkout a book. Another to allows the client to check if a book has been checkedout by a user or not. Lastly one that Returns the number of books that are checkedout currently by the user. In LibraryController added three relevant endpoints that uses the the bookService to expose the functioanlities to the client