Added review related configuration, Made the controller into three
I added a repository query where the client can find a review based on a book id and userEmail. Created a service for review, implemented the findByUserEmailAndBookId, and a query that checks if a user has already left a review on a book or not. I created a controller for review exposing the two endpoints needed, separated the other unrelated endpoints into their own corest controllers. I needed to create a dto that the client uses to send in the body of their request to create a new review. Lastly I made sure in securityConfiguration that the two new endpoints are authenticated