Added related configuration to response to currentLoans request
Added a dto that includes a book and daysleft to be used in response of call to currentLoans end point and book service. In BookService create a function that accepts userEmail and returns a list of the books the user has checkedout with related exception. The bookService makes use of a new query that that finds books based on multiple bookIds. In book controller exposed the functionality through /currentLoans endpoint. Lastly in security configuration added the endpoint to authenticated part of the app