Added history configuration
Added history entity. Added a HistoryRepository containing findBookByUserEmail that returns a Page of Pageable type later is used to be used with page and size. I enhance the returnBook function in BookService to save the related data of the book and checkout in a history entity and save it in the DB. Finally in DataRestConfig I allowed cors for /histories/**, exposed its id and disalbed its unsupported httpFunctions