Added Admin response message configuration
Added in MessageRepository findByClosed which the client app can use it to query a list of messages based on whther they are closed or not. Added an updateMessageResponse in MessageService which the admins could use to add a response to a message based on id of the message, automatically the message gets closed after this. Exposed this fuctionality in MessageController. In Security Configuration class I made suer the two new endpoints are accessible only to those users with ROLE_ADMIN role. Lastly I added the dto AdminMessageResponse that the client app uses to send a json of id and response to the server