Added admin service/controller and addBook configuraiton
Moved messageResponse endpoint to the new AdminController. Made sure in the SecurityConfiguration every endpoint staring with /api/admin can only be accesed by users with ROLE_ADMIN. Added a new DTO that the client app will use to send the newBook in that format. Finally added functionality of adding newBook to AdminService and exposed it as endpoint in AdminController