Added modify/delete book configuration
In AdminService I added two methods one for updating a book status. One for deleting a book. Exposed the functionalties in AdminController. The Delete functionality also deletes checkouts and reviews on the book being deleted for this in ReviewRepository and CheckoutRepository added a custom query on method with signature deleteAllByBookId, made sure they were annotated with @Modifying and also there were executed in a transactional context.