Added stripe payment configurations
Created an account at Stripe. Downloaded dependencies to this project. Added a variable environment in properties file that hols the stripe.key.secret that is needed for every payment. Created a Payment Entity for the table in the DB that holds outstanding late fees of users and also created a repository for it with a query of findByUserEmail. Created a PaymentInfo dto that holds currency, amount and receiptEmail of the user that is needed to be sent to Stripe to process the transaction. Added a PaymentService, injected the PaymentRepository and the variable env of the stripe.key.secret, Stripe API needs us to create a payment intent which I do, the payment intent includes the paymentMethod(in our case is card), amount and currency. I create this functionality and expose it through a new PaymentController that sends the created paymentintent to the client app in a json format. Lastly I have a pay method that the client app needs to call at the end of successfull transaction, it expects a paymentinfo, it must not be empty and the amount in it must be equal to the outstanding late-fee, then the fee will be revmoded from the user's due payment. In security Configuration I make sure the new controller endpoints are accessible only to those that are authenticated and in DataRestConfig I add it to cors.