Email configuration
Enabled Async for the whole application. Created a dto for RegisterNewUse and an entity mapped to the DB user_confirmations table. Added the mail dependency plus its properties were defined in the application.properties. Added an html to be sent to users to confirm their emails. Added an EmailService which makes use of JavaMailSender and its createMimeMessage. Add the serivce to user's service that at registration ta confirmation token will be created and the email will be sent to the user. Also Added the verifyToken service method That accepts the token and checks if its there or not. Finally in the LibraryController I expose both functionality of creating a user and verify email.